Address review comments
Remove upstreamMapper code.
Other minor fixes.
