+++ plugin/fallback/fallback_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/miekg/dns"
 
 	"fmt"
+
 	"github.com/coredns/coredns/plugin"
 	"github.com/coredns/coredns/plugin/pkg/dnstest"
 	"github.com/coredns/coredns/plugin/pkg/healthcheck"
@@ -133,7 +134,7 @@ func TestFallback(t *testing.T) {
 			handler.Next = stubNextHandler(tc.rcode, nil)
 			// add dummyUpstreams to upstream map according to the rcode field
 			for _, u := range dummyUpstreams {
-				handler.mapper.Add(u.rcode, u)
+				handler.rules[u.rcode] = u
 			}
 			proxyCreator := &testProxyCreator{t: t, expectedUpstream: tc.expectedUpstream}
 			handler.proxy = proxyCreator
@@ -165,7 +166,7 @@ func TestFallbackNotCalled(t *testing.T) {
 	handler := newFallback(nil)
 
 	// fallback only handle REFUSED
-	handler.mapper.Add(dummyRefusedUpstream.rcode, dummyRefusedUpstream)
+	handler.rules[dummyRefusedUpstream.rcode] = dummyRefusedUpstream
 
 	proxyCreator := &testProxyCreator{t: t, expectedUpstream: nil}
 	handler.proxy = proxyCreator
@@ -201,7 +202,7 @@ func TestFallbackCalledMany(t *testing.T) {
 	handler := newFallback(nil)
 	handler.Next = stubNextHandler(dns.RcodeRefused, nil)
 	// fallback only handle REFUSED
-	handler.mapper.Add(dummyRefusedUpstream.rcode, dummyRefusedUpstream)
+	handler.rules[dummyRefusedUpstream.rcode] = dummyRefusedUpstream
 	proxyCreator := &testProxyCreator{t: t, expectedUpstream: dummyRefusedUpstream}
 	handler.proxy = proxyCreator
